Updated Readme.md
Updated the Readme.md file to conclude the completion of the Minimum Viable Product.

The document is updated to outline the features included on this MVP as well as the shortcuts taken to deliver an MVP as fast as possible and other technical details and considerations.

The Readme.md also acknowledges possible improvements for the app if time allows.